Weather Actions and Reducers Removed
