+++ store/reducers/rootReducer.js
@@ -1,6 +1,5 @@
 import { combineReducers } from 'redux';
 
-import weatherReducers from './weatherReducers';
 import locationReducers from './locationReducers';
 import activityReducers from './activityReducers';
 import screenReducers from './screenReducers';
@@ -9,7 +8,6 @@ const rootReducer = combineReducers({
   activity: activityReducers,
   location: locationReducers,
   screen: screenReducers,
-  weather: weatherReducers,
 });
 
 export default rootReducer;
